September (#17)
* Create Hallas.md

Fixed size category.

* Create Oulops.md

* Update Merfolk.md

clerical fixes

* Create Mermaghu.md

* Create Mystic.md

* Update 6=Dying and Unconsciousness.md

Added fatigue when getting knocked out to represent the stress on the body.

* Rename Legendary Spells (9) .md to Legendary Spells (10) .md

* Update Creatures as a whole, adding collective nouns.

* Update 6=Dying and Unconsciousness.md

Fixed the fact that two different means of determining Incapacitation save DV were included. Will test to decide which will be better.

* Update 5=Special Actions.md

Added the ability to about face as an action or even potentially reaction because holy crap is running up on someone overpowered and the rotation mechanic is awkward in that specific case.

* Update Assassin.md

Gave the Assassin the Dire Dispose to give them a flavorful bonus for sneaking up behind someone. Plenty of narrative options with the mechanics of it.

* Rename Size Categories and Brawling to Size Categories and Brawling.md

Fixed file format.

* Update 3=Shields.md

Now have rules for multiple shields.

* Update Perdrine.md

Added Eternal Crystal perdrine power.