Update Sorcerer.md
Gave the Sorcerer the Sorcerous wave feature so they have a high level feature to look forward to that also feeds into the battle-caster aesthetic it has. Could also help out a character which doesn't get a whole lot of combat spells- particularly of a type of attack they really want to have access to.