Update Paragon.md
Lowered the initial bonus from Command: Steel Yourself from 1d6 to 1d4, but allowed the secondary +1 bonus to last for the entirety of the minute instead of simply the next save. This should make it more useful and provide a better flavor for roleplay.

Added Resolute Confidence feature because the Paragon needs to do cool stuff too.