Update Knacks.md
Added a knack for Exertive Force, which I so desperately wanted to make a class feature (especially for Assassin and Ranger) but have opted to make a knack. Those classes already look to have plenty of play options and class features and this will make it available to more characters.